Rename use-context to use-config
