+++ src/components/use-config/index.ts
